arguments cleanup
