+++ include/meta/metadata/metaclass.hpp
@@ -20,7 +20,7 @@
 #define META_METACLASS_HPP
 
 #include <assert.hpp>
-#include <meta/arguments/argument_type.hpp>
+#include <meta/arguments/packaged_arguments.hpp>
 #include <meta/forwards.hpp>
 #include <meta/meta_api.hpp>
